Make few changes in the Card, Footer components & the header =>
Add a padding top, of the height than the header, to the app div
Card component: Update html-react-parser & parse also the title of the . Set a fallback gradient background when there is no article image.
Home Carousel component: Set the  caption z-index to 1.
Footer component: Correct the svg hover effect.